safari-600.7-branch/LayoutTests/canvas/philip/tests/2d.state.saverestore.globalCompositeOperation.html

<!DOCTYPE html>
<title>Canvas test: 2d.state.saverestore.globalCompositeOperation</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

// Test that restore() undoes any modifications
var old = ctx.globalCompositeOperation;
ctx.save();
ctx.globalCompositeOperation = "copy";
ctx.restore();
_assertSame(ctx.globalCompositeOperation, old, "ctx.globalCompositeOperation", "old");

// Also test that save() doesn't modify the values
ctx.globalCompositeOperation = "copy";
old = ctx.globalCompositeOperation;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against "copy"
ctx.save();
_assertSame(ctx.globalCompositeOperation, old, "ctx.globalCompositeOperation", "old");
ctx.restore();


});
</script>
