LayoutTests/canvas/philip/tests/2d.state.saverestore.globalCompositeOperation-expected.txt

< [index] >
2d.state.saverestore.globalCompositeOperation
save()/restore() works for globalCompositeOperation
References: 2d.state.globalCompositeOperation
Actual output:
Passed
